intialization

+++ api/main.py
@@ -116,7 +116,7 @@ async def create_tables():
                 conn.execute(text("""
                     CREATE TABLE IF NOT EXISTS users (
                         id SERIAL PRIMARY KEY,
-                        telegram_id INTEGER UNIQUE NOT NULL,
+                        telegram_id BIGINT UNIQUE NOT NULL,
                         username VARCHAR,
                         first_name VARCHAR,
                         last_name VARCHAR,
